+++ packages/search/global-search/src/__tests__/unit/GlobalQuickSearch.test.tsx
@@ -4,7 +4,7 @@ import GlobalQuickSearchWithAnalytics, {
   GlobalQuickSearch,
   Props,
 } from '../../components/GlobalQuickSearch';
-import { QuickSearch } from '@atlaskit/quick-search';
+import * as AnalyticsHelper from '../../util/analytics-event-helper';
 
 const noop = () => {};
 const DEFAULT_PROPS = {
@@ -48,10 +48,166 @@ describe('GlobalQuickSearch', () => {
     const wrapper = render({ onSearch: searchMock });
 
     const onSearchInput: Function = wrapper
-      .find(QuickSearch)
+      .children()
+      .first()
       .prop('onSearchInput');
     onSearchInput({ target: { value: 'foo' } });
 
     expect(searchMock).toHaveBeenCalledWith('foo');
+  });
+
+  describe('Search result events', () => {
+    const searchSessionId = 'random-session-id';
+    let fireHighlightEventSpy;
+    let fireSearchResultSelectedEventSpy;
+    let fireAdvancedSearchSelectedEventSpy;
+    beforeEach(() => {
+      fireHighlightEventSpy = jest.spyOn(
+        AnalyticsHelper,
+        'fireHighlightedSearchResult',
+      );
+      fireSearchResultSelectedEventSpy = jest.spyOn(
+        AnalyticsHelper,
+        'fireSelectedSearchResult',
+      );
+      fireAdvancedSearchSelectedEventSpy = jest.spyOn(
+        AnalyticsHelper,
+        'fireSelectedAdvancedSearch',
+      );
+    });
+
+    afterEach(() => {
+      jest.resetAllMocks();
+    });
+
+    const deepRender = (): Function =>
+      render({ searchSessionId })
+        .dive()
+        .dive()
+        .prop('firePrivateAnalyticsEvent');
+
+    ['ArrowUp', 'ArrowDown'].forEach(key => {
+      it('should trigger highlight event', () => {
+        const firePrivateAnalyticsEvent = deepRender();
+        const eventData = {
+          resultId: 'result-id',
+          type: 'recent-result',
+          contentType: 'confluence-page',
+          sectionIndex: 2,
+          index: 11,
+          indexWithinSection: 2,
+          key,
+        };
+        // call
+        firePrivateAnalyticsEvent(
+          'atlaskit.navigation.quick-search.keyboard-controls-used',
+          eventData,
+        );
+
+        // asserts
+        expect(fireHighlightEventSpy.mock.calls.length).toBe(1);
+        expect(fireHighlightEventSpy.mock.calls[0][0]).toMatchObject(eventData);
+        expect(fireHighlightEventSpy.mock.calls[0][1]).toBe(searchSessionId);
+
+        // verify other spies are not called
+        [
+          fireAdvancedSearchSelectedEventSpy,
+          fireSearchResultSelectedEventSpy,
+        ].forEach(spy => expect(spy.mock.calls.length).toBe(0));
+      });
+    });
+
+    it('should not fire highlight event on enter', () => {
+      const firePrivateAnalyticsEvent = deepRender();
+      const eventData = {
+        resultId: 'result-id',
+        type: 'recent-result',
+        contentType: 'confluence-page',
+        sectionIndex: 2,
+        index: 11,
+        indexWithinSection: 2,
+        key: 'Enter',
+      };
+      // call
+      firePrivateAnalyticsEvent(
+        'atlaskit.navigation.quick-search.keyboard-controls-used',
+        eventData,
+      );
+
+      // verify
+      [
+        fireAdvancedSearchSelectedEventSpy,
+        fireSearchResultSelectedEventSpy,
+        fireHighlightEventSpy,
+      ].forEach(spy => expect(spy.mock.calls.length).toBe(0));
+    });
+
+    it('should fire selected search event', () => {
+      const firePrivateAnalyticsEvent = deepRender();
+      const eventData = {
+        resultId: 'result-id',
+        type: 'recent-result',
+        contentType: 'confluence-page',
+        sectionIndex: 2,
+        index: 11,
+        indexWithinSection: 2,
+        method: 'click',
+        newTab: false,
+      };
+
+      // call
+      firePrivateAnalyticsEvent(
+        'atlaskit.navigation.quick-search.submit',
+        eventData,
+      );
+
+      // assert
+      expect(fireSearchResultSelectedEventSpy.mock.calls.length).toBe(1);
+      expect(fireSearchResultSelectedEventSpy.mock.calls[0][0]).toMatchObject(
+        eventData,
+      );
+      expect(fireSearchResultSelectedEventSpy.mock.calls[0][1]).toBe(
+        searchSessionId,
+      );
+
+      // verify
+      [fireAdvancedSearchSelectedEventSpy, fireHighlightEventSpy].forEach(spy =>
+        expect(spy.mock.calls.length).toBe(0),
+      );
+    });
+
+    it('should fire advanced search event', () => {
+      const firePrivateAnalyticsEvent = deepRender();
+      const eventData = {
+        resultId: 'search_confluence',
+        type: 'recent-result',
+        contentType: 'confluence-page',
+        sectionIndex: 2,
+        index: 11,
+        indexWithinSection: 2,
+        method: 'click',
+        newTab: false,
+      };
+
+      // call
+      firePrivateAnalyticsEvent(
+        'atlaskit.navigation.quick-search.submit',
+        eventData,
+      );
+
+      // assert
+      expect(fireAdvancedSearchSelectedEventSpy.mock.calls.length).toBe(1);
+      expect(fireAdvancedSearchSelectedEventSpy.mock.calls[0][0]).toMatchObject(
+        eventData,
+      );
+      expect(fireAdvancedSearchSelectedEventSpy.mock.calls[0][1]).toBe(
+        searchSessionId,
+      );
+
+      // verify
+      [fireSearchResultSelectedEventSpy, fireHighlightEventSpy].forEach(spy =>
+        expect(spy.mock.calls.length).toBe(0),
+      );
+    });
   });
 });