+++ packages/search/global-search/src/components/GlobalQuickSearch.tsx
@@ -3,18 +3,31 @@ import * as React from 'react';
 import * as debounce from 'lodash.debounce';
 import { QuickSearch } from '@atlaskit/quick-search';
 import { LinkComponent } from './GlobalQuickSearchWrapper';
-import { GasPayload } from '@atlaskit/analytics-gas-types';
 import {
   withAnalyticsEvents,
   AnalyticsContext,
 } from '@atlaskit/analytics-next';
 
+import { withAnalytics } from '@atlaskit/analytics';
+
 import {
-  DEFAULT_GAS_ATTRIBUTES,
-  DEFAULT_GAS_SOURCE,
-  DEFAULT_GAS_CHANNEL,
-  sanitizeSearchQuery,
-} from '../util/analytics-util';
+  fireSelectedSearchResult,
+  fireHighlightedSearchResult,
+  SelectedSearchResultEvent,
+  AdvancedSearchSelectedEvent,
+  KeyboardControlEvent,
+  SearchResultEvent,
+  fireSelectedAdvancedSearch,
+  fireTextEnteredEvent,
+  fireDismissedEvent,
+} from '../util/analytics-event-helper';
+
+import { CreateAnalyticsEventFn } from './analytics/types';
+import { isAdvancedSearchResult } from './SearchResultsUtil';
+
+const ATLASKIT_QUICKSEARCH_NS = 'atlaskit.navigation.quick-search';
+const QS_ANALYTICS_EV_KB_CTRLS_USED = `${ATLASKIT_QUICKSEARCH_NS}.keyboard-controls-used`;
+const QS_ANALYTICS_EV_SUBMIT = `${ATLASKIT_QUICKSEARCH_NS}.submit`;
 
 export interface Props {
   onMount();
@@ -27,7 +40,7 @@ export interface Props {
   searchSessionId: string;
   children: React.ReactNode;
   linkComponent?: LinkComponent;
-  createAnalyticsEvent?: Function;
+  createAnalyticsEvent?: CreateAnalyticsEventFn;
 }
 
 /**
@@ -48,52 +61,60 @@ export class GlobalQuickSearch extends React.Component<Props> {
   debouncedSearch = debounce(this.doSearch, 350);
 
   doSearch(query: string) {
-    this.props.onSearch(query);
-
-    if (this.props.createAnalyticsEvent) {
-      const sanitizedQuery = sanitizeSearchQuery(query);
-      const event = this.props.createAnalyticsEvent();
-      const searchSessionId = this.props.searchSessionId;
-      const payload: GasPayload = {
-        action: 'entered',
-        actionSubject: 'text',
-        eventType: 'track',
-        source: DEFAULT_GAS_SOURCE,
-        attributes: {
-          queryId: null,
-          queryVersion: this.queryVersion,
-          queryLength: sanitizedQuery.length,
-          wordCount:
-            sanitizedQuery.length > 0 ? sanitizedQuery.split(/\s/).length : 0,
-          ...DEFAULT_GAS_ATTRIBUTES,
-          searchSessionId: searchSessionId,
-        },
-      };
-      event.update(payload).fire(DEFAULT_GAS_CHANNEL);
-    }
-
+    const { onSearch, searchSessionId, createAnalyticsEvent } = this.props;
+    onSearch(query);
+    fireTextEnteredEvent(
+      query,
+      searchSessionId,
+      this.queryVersion,
+      createAnalyticsEvent,
+    );
     this.queryVersion++;
   }
 
-  componentWillUnmount() {
-    const { createAnalyticsEvent } = this.props;
-    if (createAnalyticsEvent) {
-      // Note: This analytics event is currently missing the
-      // trigger attribute, to indicate _how_ the drawer was dismissed.
-      // as well as the correct actionSubjectId.
-      const event = createAnalyticsEvent();
-      const payload: GasPayload = {
-        action: 'dismissed',
-        actionSubject: 'globalSearchDrawer',
-        source: DEFAULT_GAS_SOURCE,
-        eventType: 'ui',
-        attributes: {
-          searchSessionId: this.props.searchSessionId,
-          ...DEFAULT_GAS_ATTRIBUTES,
-        },
-      };
-      event.update(payload).fire(DEFAULT_GAS_CHANNEL);
+  fireSearchResultSelectedEvent = (eventData: SearchResultEvent) => {
+    const { createAnalyticsEvent, searchSessionId, query } = this.props;
+    if (isAdvancedSearchResult(eventData.resultId)) {
+      fireSelectedAdvancedSearch(
+        {
+          ...eventData,
+          query,
+          queryVersion: this.queryVersion,
+        } as AdvancedSearchSelectedEvent,
+        searchSessionId,
+        createAnalyticsEvent,
+      );
+    } else {
+      fireSelectedSearchResult(
+        eventData as SelectedSearchResultEvent,
+        searchSessionId,
+        createAnalyticsEvent,
+      );
     }
+  };
+
+  fireSearchResultEvents = (
+    eventName: string,
+    eventData: SearchResultEvent,
+  ) => {
+    const { createAnalyticsEvent, searchSessionId } = this.props;
+    if (eventName === QS_ANALYTICS_EV_SUBMIT) {
+      this.fireSearchResultSelectedEvent(eventData);
+    } else if (eventName === QS_ANALYTICS_EV_KB_CTRLS_USED) {
+      const data = eventData as KeyboardControlEvent;
+      if (data.key === 'ArrowDown' || data.key === 'ArrowUp') {
+        fireHighlightedSearchResult(
+          data,
+          searchSessionId,
+          createAnalyticsEvent,
+        );
+      }
+    }
+  };
+
+  componentWillUnmount() {
+    const { createAnalyticsEvent, searchSessionId } = this.props;
+    fireDismissedEvent(searchSessionId, createAnalyticsEvent);
   }
 
   render() {
@@ -106,9 +127,17 @@ export class GlobalQuickSearch extends React.Component<Props> {
       onSearchSubmit,
     } = this.props;
 
+    const QuickSearchWithAnalytics = withAnalytics(
+      QuickSearch,
+      {
+        firePrivateAnalyticsEvent: this.fireSearchResultEvents,
+      },
+      {},
+    );
+
     return (
       <AnalyticsContext data={{ searchSessionId: this.props.searchSessionId }}>
-        <QuickSearch
+        <QuickSearchWithAnalytics
           isLoading={isLoading}
           onSearchInput={this.handleSearchInput}
           placeholder={placeholder}
@@ -117,7 +146,7 @@ export class GlobalQuickSearch extends React.Component<Props> {
           onSearchSubmit={onSearchSubmit}
         >
           {children}
-        </QuickSearch>
+        </QuickSearchWithAnalytics>
       </AnalyticsContext>
     );
   }