+++ packages/search/global-search/CHANGELOG.md
@@ -1,5 +1,10 @@
 # @atlaskit/global-search
 
+## 5.0.1
+- [patch] Add extra analytics event for highlight and selection of a search result [12e79bf](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/12e79bf)
+- [patch] Updated dependencies [12e79bf](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/12e79bf)
+  - @atlaskit/quick-search@4.0.1
+
 ## 5.0.0
 
 - [major] Updates to React ^16.4.0 [7edb866](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/7edb866)